==> parking_demand_competion-main/01_competition_firstmodel.ipynb <==
### 대회 첫 모델 만들기
 * 데이터 살펴보기

import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.linear_model import LinearRegression

import pandas as pd

train = pd.read_csv("../data/parking_demand/train.csv")
test = pd.read_csv("../data/parking_demand/test.csv")
sub = pd.read_csv("../data/parking_demand/sample_submission.csv")
age = pd.read_csv("../data/parking_demand/age_gender_info.csv")

train.shape, test.shape, sub.shape, age.shape

train.info()

test.info()

### 결측치가 얼마나 될까?

train.isna().sum()

test.isna().sum()

* 임대보증금, 임대료, 지하철역수, 버스정류장수(train only), 자격유형(test only)

train.head()

test.head()

train.columns

sel = ['총세대수']
X_train = train[sel]
# X_train = train['총세대수']
print(X_train.shape)
X_test = test[sel]
y_train = train['등록차량수']

from sklearn.linear_model import LinearRegression

model = LinearRegression()
model.fit(X_train, y_train)
pred = model.predict(X_test)
pred

test.shape, sub.shape

test.columns

len( test['단지코드'].unique() )

test['등록차량수'] = pred

test.groupby('단지코드')['등록차량수'].mean()

import numpy as np

test['코드별차량수평균'] = test.groupby('단지코드')['등록차량수'].transform(np.mean)
test.head(10)

test.drop_duplicates(['단지코드'], keep='first')

test_new = test.drop_duplicates(['단지코드'], keep='first').reset_index()
test_new

sub_df = test_new[ ['단지코드', '코드별차량수평균']]
sub_df.columns = ['code', 'num']
sub_df

sub_df.to_csv('baseline_0712.csv', index=False)

import os
os.listdir(os.getcwd())
