==> parking_demand_competion-main/03_second_linear_model-Copy2.ipynb <==
### 학습 목표 
 * 오류 처리된(2,3번 해결) 데이터를 활용한다.
 * 데이터 EDA 및 시각화를 통해 데이터를 이해하고 기본 모델을 만들어본다.
 * 모델을 제출해 본다.

* 데이콘 대회 : https://dacon.io/competitions/official/235745/overview/description
* https://dacon.io/competitions/official/235745/talkboard/403708?page=1&dtype=recent

### 01. 데이터 불러오기 및 확인

import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.linear_model import LinearRegression

import pandas as pd

train = pd.read_csv("../data/parking_demand/train_df_errno.csv")
test = pd.read_csv("../data/parking_demand/test_df.csv")
sub = pd.read_csv("../data/parking_demand/sample_submission.csv")
age = pd.read_csv("../data/parking_demand/age_gender_info.csv")

train.shape, test.shape, sub.shape, age.shape

train.columns

train.columns = ['단지코드', '총세대수', '임대건물구분', '지역', '공급유형', '전용면적', '전용면적별세대수', '공가수',
       '자격유형', '임대보증금', '임대료', '10분내지하철수',
       '10분내버스정류장수', '단지내주차면수', '등록차량수']

test.columns = ['단지코드', '총세대수', '임대건물구분', '지역', '공급유형', '전용면적', '전용면적별세대수', '공가수',
       '자격유형', '임대보증금', '임대료', '10분내지하철수',
       '10분내버스정류장수', '단지내주차면수']


### 오류 데이터가 없는지 확인

train.loc[ ((train['단지코드']=='C2675') | 
           (train['단지코드']=='C2335') |
           (train['단지코드']=='C1327') ) , :]

test.loc[ ((test['단지코드']=='C2675') | 
           (test['단지코드']=='C2335') |
           (test['단지코드']=='C1327') ) , :]

### 오류 데이터 처리 확인

* ※ 동일한 단지에 코드가 2개로 부여된 단지 코드 (3쌍) : ['C2085', 'C1397'], ['C2431', 'C1649'], ['C1036', 'C2675'] 

* - (참고 사항) 주차면수는 하나의 단지임을 전제로 산정된 것이고 총세대수는 두 개 단지의 합계입니다. 다만 등록차량대수는 ['C2085', 'C1397'] 단지의 경우 동일 수치

# C2085, C1397 -> N2085
train.loc[ train['단지코드']=='N2085', : ].shape

### 오류 코드 변경
 * C2431, C1649의 총세대수를 1047로 변경
 * C2431, C1649의 등록차량대수를 1214로 변경
 * C2431, C1649의 단지코드를 N2431로 변경
 

# C2431, C1649 -> N2431
print( train.loc[ train['단지코드']=='N2431', : ].shape )
train.loc[ train['단지코드']=='N2431', : ]

### 오류 코드 변경
 * C1036의 총세대수를 1243로 변경
 * C1036의 단지코드를 N1036로 변경

# C2085, C1397 -> N2085
train.loc[ train['단지코드']=='N1036', : ].shape

### 오류 3
```
3. 단지코드 등 기입 실수로 데이터 정제 과정에서 매칭 오류 발생  
 - (오류 내용) 단지코드 등 기입 실수로 총세대수가 주차면수에 비해 과하게 많거나 적은 경우가 발생하였고, 점검 결과 일부 데이터의 단지코드, 총세대수, 주차면수 등에서 오류가 검출되었습니다.
 - (발생 원인) 원천데이터 수집 과정에서 단지 코드 등이 잘못 기입되었고 이를 인지하지 못한 채 데이터 정제를 하여 오류가 발생하였습니다.
 - (관련 데이터) 아래와 같이 총 9개 단지에서 같은 문제가 확인되었습니다. 
※ 실수가 발생한 단지 코드 (9개 단지) : ['C2335', 'C1327', 'C1095', 'C2051', 'C1218', 'C1894', 'C2483', 'C1502', 'C1988']
 - C2335, C1327 단지는 테스트셋, 나머지는 트레인셋 입니다.
```

### 오류 처리
 * train 데이터 셋에 오류 발생 코드를 ERR04로 변경 후, 데이터 셋을 두개로 분리

train.loc[ train['단지코드'].str.contains('ERR'), :].shape

### 02. 결측치를 처리(1)

train.isnull().sum()

test.isnull().sum()

###  자격유형(test) 결측치 처리

train['지역'].unique()

test['지역'].unique()

test.loc[test['자격유형'].isnull()]

grouped = test.groupby(['단지코드', '임대건물구분', '지역','공급유형'])
group1 = grouped.get_group( ('C2411', '아파트', '경상남도', '국민임대')  )
group2 = grouped.get_group( ('C2253', '아파트', '강원도', '영구임대')  )
group2

test.loc[ 196, "자격유형"] = 'A'
test.loc[ 258, "자격유형"] = 'C'

train.head(3)

test.head(3)

print(train.자격유형.unique())
print(test.자격유형.unique())

mapping = { 'A':1, 'B':2, 'C':3, 'D':4, 'E':5, 
            'F':6, 'G':7, 'H':8, 'I':9, 'J':10, 
            'K':11, 'L':12, 'M':13, 'N':14, 'O':15  }

train['자격유형'] =train['자격유형'].map(mapping).astype(int)
test['자격유형'] =test['자격유형'].map(mapping).astype(int)

train.head(3)

test.head(3)

print(train.공급유형.unique())
print(test.공급유형.unique())

train.columns

### 10분내버스정류장수 (tr) 결측치 처리,

train.isnull().sum()

train.loc[ train['10분내버스정류장수'].isnull(), :]

grouped = train.groupby(['임대건물구분', '지역','공급유형', '자격유형'])
group1 = grouped.get_group( ('아파트', '경상남도', '공공임대(10년)', 1)  )
group1

# 데이터 확인 후, 임의 처리 4
train.loc[ train['10분내버스정류장수'].isnull(), "10분내버스정류장수"] = 4

train.loc[ train['10분내버스정류장수'].isnull(), :]

train.corr()['등록차량수']

from sklearn.linear_model import LinearRegression
import numpy as np
from sklearn.model_selection import train_test_split

sel = ['총세대수', '전용면적', '전용면적별세대수', '공가수',  '단지내주차면수', '자격유형']

X = train[sel]
y = train['등록차량수']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, 
                                                   random_state=0)

model = LinearRegression()
model.fit(X_train, y_train)
pred = model.predict(X_test)

print("학습(score) :", model.score(X_train, y_train) ) # 결정계수
print("테스트(score) :", model.score(X_test, y_test) ) # 결정계수

# mae
np.mean( np.abs(y_test - pred)  )

# mse
mse_val = np.mean( (y_test - pred)**2  )
mse_val

# rmse
np.sqrt(  mse_val )

test['등록차량수'] = pred
test['단지별차량수평균'] = test.groupby("단지코드")['등록차량수'].transform(np.mean)
test_new = test.drop_duplicates(['단지코드'], keep='first').reset_index()
test_new

test_X = test[sel]

add_dat = {'code':['C2675', 'C2335', 'C1327'],
           'num':['0', '0', '0']}
add_df = pd.DataFrame(add_dat)
add_df

sub_df = test_new[ ['단지코드', '단지별차량수평균']]
sub_df.columns = ['code', 'num']
sub_df = pd.concat([sub_df, add_df]).reset_index()
sub_df = sub_df.drop(['index'], axis=1)
sub_df

sub_df.to_csv('second_rf_0712.csv', index=False)
sub_df.head()

import os
os.listdir(os.getcwd())

### 점수 : 138.65787
